finished fixing manual data, added test to make sure manual data remains consistent
