+++ contacts/manual_data/U.S. DOL.yaml
@@ -7,7 +7,7 @@ departments:
   top_level: true
 - name: Employment & Training Administration
   top_level: true
-- name: Job Corps
+- name: Jobs Corps
   top_level: true
 - name: Mine Safety & Health Administration
   top_level: true
